as per the discussion on kde-core-devel, add a KDE_SESSION_VERSION env var that holds the value of the major # of the running desktop, allowing one to run kde${KDE_SESSION_VERSION}-config and get the proper kde?-config.
i considered using cmake's configure_file to process this and add the 4
directly, but since this is only getting changed every 5 years or so and
it would mean adding another cmake variable to the build, etc... i took
the simple way out.

if there are objections to this being managed "by hand" i'll try and
find the energy to automate it.

svn path=/trunk/KDE/kdebase/workspace/; revision=630924