auto-add default domain to static domain list
this doesn't make much sense if a static domain list is provided anyway,
as then the default domain should be listed there already.
however, if the user relies on the dynamic domain listing, needing to
explicitly put the default domain into the otherwise empty static domain
list is a bit counterintuitive.
the alternative would have been making the default domain work with the
dynamically fetched domain list, but that kind of makes no sense, as we
apparently know that domain, so it can be made static just as well.

svn path=/trunk/KDE/kdebase/workspace/; revision=1184454