Revert "Remove the FindQJSON cmake module, as QJSON provides a perfectly good QJSONConfig.cmake which is more accurate"
As discussed with Ben on #kde-devel, QJSONConfig.cmake was added to qjson
after the latest stable release (0.7.1), so we cannot rely on that.

As the build failed only on build.kde.org and the logs have been lost, I
will tentatively write a new FindQJSON.cmake that hopefully avoids the
previous build failures.

This reverts commit ce878021b9a1edd13f9b3142e15f68d4bab5c944.